добавлен декорер,подредактированы функци в todo, все работают но с багом (не получилось сделать ф-цию пересчета нумерации), выделен todo в отдельную папку
