--- my_todo.json
@@ -1,1 +1,0 @@
-{"1": "tuz", "2": "bus", "3": "pop", "4": "rap13", "5": "rock1", "6": "bus88", "8": "1"}
