+++ my_todo/hw_4_note.py
@@ -1,6 +1,6 @@
 print ('Записная приветствует Вас')
 notebook = {1:'Milk', 2:'Bread', 3:'Coffee', 4:'Sausage', 5: 'Salt', 6:'Fish'}
-step = 10
+step = 0
 while step:
     step = input ('Ваше действие: \n1. Просмотреть список. \n2. добавить задачу \n3. удалить задачу. \n4. редактировать задачу. \n5. отметить задачу. \n0. ВЫЙТИ \n:')
     if step == '1':
